Replace product form
