+++ templates/skeleton/app/components/ProductForm.tsx
@@ -1,35 +1,91 @@
-import {Link} from '@remix-run/react';
-import {type VariantOption, VariantSelector} from '@shopify/hydrogen';
+import {Link, useNavigate} from '@remix-run/react';
+import {type MappedProductOptions} from '@shopify/hydrogen';
 import type {
-  ProductFragment,
-  ProductVariantFragment,
-} from 'storefrontapi.generated';
-import {AddToCartButton} from '~/components/AddToCartButton';
-import {useAside} from '~/components/Aside';
+  Maybe,
+  ProductOptionValueSwatch,
+} from '@shopify/hydrogen/storefront-api-types';
+import {AddToCartButton} from './AddToCartButton';
+import {useAside} from './Aside';
+import type {ProductFragment} from 'storefrontapi.generated';
 
 export function ProductForm({
-  product,
+  productOptions,
   selectedVariant,
-  variants,
 }: {
-  product: ProductFragment;
+  productOptions: MappedProductOptions[];
   selectedVariant: ProductFragment['selectedOrFirstAvailableVariant'];
-  variants: Array<ProductVariantFragment>;
 }) {
+  const navigate = useNavigate();
   const {open} = useAside();
   return (
-    <div className="product-form">
-      <VariantSelector
-        handle={product.handle}
-        options={product.options.filter(
-          (option) => option.optionValues.length > 1,
-        )}
-        variants={variants}
-        selectedVariant={selectedVariant}
-      >
-        {({option}) => <ProductOptions key={option.name} option={option} />}
-      </VariantSelector>
-      <br />
+    <>
+      {productOptions.map((option) => (
+        <div className="product-options" key={option.name}>
+          <h5>{option.name}</h5>
+          <div className="product-options-grid">
+            {option.optionValues.map((value) => {
+              const {
+                name,
+                handle,
+                variantUriQuery,
+                selected,
+                available,
+                exists,
+                isDifferentProduct,
+                swatch,
+              } = value;
+
+              if (isDifferentProduct) {
+                return (
+                  <Link
+                    className="product-options-item"
+                    key={option.name + name}
+                    prefetch="intent"
+                    preventScrollReset
+                    replace
+                    to={`/products/${handle}?${variantUriQuery}`}
+                    style={{
+                      border: selected
+                        ? '1px solid black'
+                        : '1px solid transparent',
+                      opacity: available ? 1 : 0.3,
+                    }}
+                  >
+                    <ProductOptionSwatch swatch={swatch} name={name} />
+                  </Link>
+                );
+              } else {
+                return (
+                  <button
+                    type="button"
+                    className={`product-options-item${
+                      exists && !selected ? ' link' : ''
+                    }`}
+                    key={option.name + name}
+                    style={{
+                      border: selected
+                        ? '1px solid black'
+                        : '1px solid transparent',
+                      opacity: available ? 1 : 0.3,
+                    }}
+                    disabled={!exists}
+                    onClick={() => {
+                      if (!selected) {
+                        navigate(`?${variantUriQuery}`, {
+                          replace: true,
+                        });
+                      }
+                    }}
+                  >
+                    <ProductOptionSwatch swatch={swatch} name={name} />
+                  </button>
+                );
+              }
+            })}
+          </div>
+          <br />
+        </div>
+      ))}
       <AddToCartButton
         disabled={!selectedVariant || !selectedVariant.availableForSale}
         onClick={() => {
@@ -49,35 +105,31 @@ export function ProductForm({
       >
         {selectedVariant?.availableForSale ? 'Add to cart' : 'Sold out'}
       </AddToCartButton>
-    </div>
+    </>
   );
 }
 
-function ProductOptions({option}: {option: VariantOption}) {
+function ProductOptionSwatch({
+  swatch,
+  name,
+}: {
+  swatch?: Maybe<ProductOptionValueSwatch> | undefined;
+  name: string;
+}) {
+  const image = swatch?.image?.previewImage?.url;
+  const color = swatch?.color;
+
+  if (!image && !color) return name;
+
   return (
-    <div className="product-options" key={option.name}>
-      <h5>{option.name}</h5>
-      <div className="product-options-grid">
-        {option.values.map(({value, isAvailable, isActive, to}) => {
-          return (
-            <Link
-              className="product-options-item"
-              key={option.name + value}
-              prefetch="intent"
-              preventScrollReset
-              replace
-              to={to}
-              style={{
-                border: isActive ? '1px solid black' : '1px solid transparent',
-                opacity: isAvailable ? 1 : 0.3,
-              }}
-            >
-              {value}
-            </Link>
-          );
-        })}
-      </div>
-      <br />
+    <div
+      aria-label={name}
+      className="product-option-label-swatch"
+      style={{
+        backgroundColor: color || 'transparent',
+      }}
+    >
+      {!!image && <img src={image} alt={name} />}
     </div>
   );
 }
